Update _index.en.md
Modification to additional identifiers definition. I think it's of value to define this field in a broader context to the traditional dwc:otherCatalogNumbers definition, and thus the reason to not use the same terminology. Yet, within the Darwin Core exports/imports, that is the definition that these identifiers map best. At some point, we will be adding an identifier extension file to the DwC-A exports.